stream_processor/pyproject.toml
[tool.poetry]
name = "stream-processor"
version = "0.1.0"
description = ""
authors = ["Mihajlo Madic <mihajlo.madic@elfak.rs>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyspark = "^3.5.1"
google-cloud-bigquery = "^3.23.1"
pyarrow = "^16.1.0"
pandas = "^2.2.2"
google-cloud-bigquery-storage = "^2.25.0"
python-dotenv = "^1.0.1"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
